Rename middleware folder to routes
