+++ apps/photo-storage/routes/assertBucket.js
