+++ apps/photo-storage/routes/delete.js
